refactor(pi): own failure history extraction in harness
