+++ tests/unit/harness/test_pi_failure.py
@@ -6,12 +6,11 @@ import json
 
 import pytest
 
-from meridian.lib.harness.extractors.pi import PI_EXTRACTOR as _PI_EXTRACTOR
-from meridian.lib.harness.pi_failure import compact_pi_failure_output
+from meridian.lib.harness.pi_failure import (
+    compact_pi_failure_output,
+    extract_pi_failure_from_history,
+)
 from meridian.lib.launch.errors import should_retry
-from meridian.lib.launch.report import extract_pi_failure_from_history
-
-_ = _PI_EXTRACTOR  # Initialize harness modules before importing launch.report in isolation.
 
 
 def test_extract_pi_failure_from_prompt_rejection() -> None:
